requirements.txt
opencv-python
tensorflow==2.1
pathlib
numpy
matplotlib
Pillow
gym
pyserial